Deleted Dump extension, refactored LoadFromDump
